Map remote filenames to local for remote debugging.
Store the launch arguments for later use
Improve remote debugging by allowing regex file name mappings in a
'tolocal' block in the launch arguments:

  "tolocal": [ { "<exp>": "<subSpec>", ... }, ... ]

The arguments 'exp' and 'subSpec' are as supplied to TCL's regsub
command and named accordingly. The format of 'tolocal' is a list of
objects whose .  The keys of the objects are regular expressions ('exp')
and the values are the replacements ('subSpec'). The first mapping which
matches is used. The expressions are evaluated in the sequence they
appear in the list. Note that values within an object are not sequenced,
but allowed to make reading/writing simpler where the expressions are
not sequence dependent.

Example from vimspector:

    "t: {
        "adapter": "tclpro",
        "remote-request": "launch",
        "remote-cmdLine": [ "run-test", "testing" ],
        "configuration": {
            "request": "attach",
            "pauseOnEntry": true,
            "host": "${host}",
            "port": "${port}",
            "tolocal": [
                {
                    "^.*/runtime/(testing|overlay)/(tcl|scripts)/testing_(.*)":
                        "${workspaceRoot}/src/\\2/\\3",
                    "^.*/runtime/(test-test-test|overlay)/(tcl|scripts)/test-test-test_(.*)":
                        "${workspaceRoot}/src/TEST/\\2/\\3"
                },
                {
                    "^.*/runtime/dependency/": "/path/to/dependency/",
                }
            ]
        }
    }