Support list, dict and array display
For watch windows we just can't know (or guess) what the type of the
returned value should be interpretted as, so we instroduce type
suffixes:

* <expr>,l - treat the result as a list
* <expr>,d - treat the result as a dict
* <expr>,s - treat the result as a scalar

When we get the eval response, use a new list to store the results and
return variablesReferences as indices into this list. There's oinly one
namespace for variablesReferences, so use an arbitrary number for now.

For the scopes window we can know that a name is an array, so in that
case we knnow that the debugger will return a name/val and we can re-use
our dict display code.

For scalar values when displaying the initial value we truncate the
string and offser expansion. The expanded items include:

* Value - the full value
* As list... - the vaule interpreted as a list (as in ,l)
* As dict... - the vaule interpreted as a dict (as in ,l)

This gives a nice discoverable way of doing this and also works for the
scopes window!